man kan nu se forum selvom man er 'logout'

+++ JTM/Forum/Category.aspx.cs
@@ -38,14 +38,17 @@ public partial class Forum_Topic : System.Web.UI.Page
                 html += "<tr>";
                 html += "<td class='leftpart'>";
                 html += "<h3><a href='Topic.aspx?id=" + getTop[i][0] + "'>" + getTop[i][1] + "</a>";
-                if (Request.Cookies["forumcookie"]["userlevel"] == "0")
+                if (Request.Cookies["forumcookie"] != null)
                 {
-                    html += " <a class='item' href='Delete_Topic.aspx?id='" + getTop[i][0] + "'>Slet tråd</a>";
-                    html += "<a class='item' href='Lock_Topic.aspx?id='" + getTop[i][0] + "'>Lås tråd</a>";
-                }
-                else
-                {
-                    html += "</h3>";
+                    if (Request.Cookies["forumcookie"]["userlevel"] == "0")
+                    {
+                        html += " <a class='item' href='Delete_Topic.aspx?id='" + getTop[i][0] + "'>Slet tråd</a>";
+                        html += "<a class='item' href='Lock_Topic.aspx?id='" + getTop[i][0] + "'>Lås tråd</a>";
+                    }
+                    else
+                    {
+                        html += "</h3>";
+                    }
                 }
                 html += "</td>";
                 html += "<td class='rightpart'>";
